Switch frontend tests to Weaver
