+++ build.sbt
@@ -10,7 +10,7 @@ val V = new {
   val decline          = "1.3.0"
   val organiseImports  = "0.4.0"
   val betterMonadicFor = "0.3.1"
-  val utest            = "0.7.5"
+  val weaver           = "0.5.0"
 }
 
 val Dependencies = new {
@@ -22,8 +22,7 @@ val Dependencies = new {
   lazy val frontend = Seq(
     libraryDependencies ++=
       sttpModules.map("com.softwaremill.sttp.client" %%% _         % V.sttp) ++
-        Seq("com.raquo"                              %%% "laminar" % V.laminar) ++
-        Seq("com.lihaoyi" %%% "utest" % V.utest % Test)
+        Seq("com.raquo"                              %%% "laminar" % V.laminar)
   )
 
   lazy val backend = Seq(
@@ -32,10 +31,14 @@ val Dependencies = new {
         Seq("com.monovore"           %% "decline" % V.decline)
   )
 
-  lazy val shared = new {
-    val js  = libraryDependencies += "io.circe" %%% "circe-generic" % V.circe
-    val jvm = libraryDependencies += "io.circe"  %% "circe-generic" % V.circe
-  }
+  lazy val shared = Def.settings(
+    libraryDependencies += "io.circe" %%% "circe-generic" % V.circe
+  )
+
+  lazy val tests = Def.settings(
+    libraryDependencies += "com.disneystreaming" %%% "weaver-framework" % V.weaver % Test,
+    testFrameworks += new TestFramework("weaver.framework.TestFramework")
+  )
 }
 
 inThisBuild(
@@ -56,8 +59,7 @@ lazy val frontend = (project in file("modules/frontend"))
   .settings(scalaJSUseMainModuleInitializer := true)
   .settings(
     Dependencies.frontend,
-    Dependencies.shared.js,
-    testFrameworks += new TestFramework("utest.runner.Framework"),
+    Dependencies.tests,
     jsEnv in Test := new org.scalajs.jsenv.jsdomnodejs.JSDOMNodeJSEnv()
   )
   .settings(commonBuildSettings)
@@ -65,6 +67,7 @@ lazy val frontend = (project in file("modules/frontend"))
 lazy val backend = (project in file("modules/backend"))
   .dependsOn(shared.jvm)
   .settings(Dependencies.backend)
+  .settings(Dependencies.tests)
   .settings(commonBuildSettings)
   .enablePlugins(JavaAppPackaging)
   .enablePlugins(DockerPlugin)
@@ -83,8 +86,8 @@ lazy val backend = (project in file("modules/backend"))
 lazy val shared = crossProject(JSPlatform, JVMPlatform)
   .crossType(CrossType.Pure)
   .in(file("modules/shared"))
-  .jvmSettings(Dependencies.shared.jvm)
-  .jsSettings(Dependencies.shared.js)
+  .jvmSettings(Dependencies.shared)
+  .jsSettings(Dependencies.shared)
   .jsSettings(commonBuildSettings)
   .jvmSettings(commonBuildSettings)
